refactor: central python project
