+++ doc/documentation/src/tutorial_templates/tutorial_poisson.rst.j2
@@ -253,6 +253,6 @@ Thus, the error between the |FOURC| and deal.II solutions is within this range, 
    Error between diffusion and thermal conduction,
    and between deal.II and thermal conduction along the line at :math:`y=0`.
 
-Here is a :download:`ParaView Python script </_assets/tutorial_poisson_pvscript.py>`
+Here is a `ParaView Python script <https://github.com/4C-multiphysics/4C/blob/main/utilities/four_c_python/src/four_c_documentation/tutorial_poisson_pvscript.py>`
 that generates the figure for the structure loaded in ParaView (visible in Paraview's current RenderView).
