+++ utilities/four_c_python/README.md
@@ -1,7 +1,8 @@
 A Python package that bundles functionality used with/in 4C.
 
 It contains 4 Python packages
+- `four_c_building_testing`: Tools related to building and testing
 - `four_c_common_utils`: Utilities needed by other packages
 - `four_c_development`: Tools required for development
+- `four_c_documentation`: Tools and scripts for the documentation
 - `four_c_metadata`: Tools related to input files and its metadata
-- `four_c_building_testing`: Tools related to building and testing
