+++ utilities/four_c_python/pyproject.toml
@@ -22,7 +22,11 @@ check-includes-for-cycles = "four_c_development.check_includes_for_cycles:main"
 check-header-guards = "four_c_development.check_header_guards:main"
 check-test-files = "four_c_development.check_test_files:main"
 check-preprocessor = "four_c_development.check_preprocessor:main"
+check-python-files = "four_c_development.check_python_files:main"
 check-workflow-dependencies-hash = "four_c_development.check_workflow_dependencies_hash:main"
+chunk-test-suite = "four_c_building_testing.chunk_test_suite:main"
+clang-tidy-filter-config = "four_c_development.clang_tidy_filter_config:main"
+clang-tidy-filter-output = "four_c_development.clang_tidy_filter_output:main"
 create-schema-files = "four_c_metadata.json_schema:cli"
 validate-test-files = "four_c_metadata.validate_test_input_files:cli"
 diff-with-tolerance = "four_c_building_testing.diff_with_tolerance:cli"
